# internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for reading and writing Bear notes via the hub.",
        "title": "Salmon Hub API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/attachments/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Downloads the attachment file by ID. Returns the file with Content-Disposition header.",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Attachments"
                ],
                "summary": "Get an attachment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Attachment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/notes/{noteID}/attachments": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Uploads a file and attaches it to a note. Max file size is 5 MB. Requires Idempotency-Key header.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notes"
                ],
                "summary": "Attach a file to a note",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Note ID",
                        "name": "noteID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Idempotency key for deduplication",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "File to attach",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/models.WriteQueueItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/notes/{noteID}/backlinks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all notes that link to the specified note.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backlinks"
                ],
                "summary": "List backlinks for a note",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Note ID",
                        "name": "noteID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Backlink"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "descriptive error message"
                }
            }
        },
...
        "models.Backlink": {
            "type": "object",
            "properties": {
                "bear_id": {
                    "type": "string"
                },
                "bear_raw": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "linked_by_id": {
                    "description": "note that contains the link",
                    "type": "string"
                },
                "linking_to_id": {
                    "description": "note being linked to",
                    "type": "string"
                },
                "location": {
                    "type": "integer"
                },
                "modified_at": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
        "models.WriteQueueItem": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "create | update | add_tag | trash | add_file | archive | rename_tag | delete_tag",
                    "type": "string"
                },
                "applied_at": {
                    "type": "string"
                },
                "consumer_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "lease_until": {
                    "type": "string"
                },
                "note_id": {
                    "type": "string"
                },
                "note_sync_status": {
                    "description": "NoteSyncStatus is populated during lease to indicate the associated note's sync_status.\nNot stored in DB — filled by JOIN during LeaseQueueItems.",
                    "type": "string"
                },
                "payload": {
                    "description": "JSON with operation data",
                    "type": "string"
                },
                "processing_by": {
                    "type": "string"
                },
                "status": {
                    "description": "pending | processing | applied | failed",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Consumer Bearer token. Format: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}